+++ iowmongotools/cluster.py
@@ -58,14 +58,14 @@ class Cluster(object):
         """
         self._declared_config = cluster_config
         self._api = pymongo.MongoClient(['mongodb://%s' % mongos for mongos in cluster_config['mongos']], connect=False)
-        self._name = name
+        self.name = name
 
     def generate_commands(self):
         """ :returns dict of lists of commands """
         sharded_dbs = (db for db, params in self._declared_config['databases'].items() if params['partitioned'])
         commands = {
             'add_shards': [app.Command(self._api.admin.command, {'addshard': shard, 'name': shard.split('.')[0]},
-                                       'adding shard %s to %s' % (shard, self._name)) for shard in
+                                       'adding shard %s to %s' % (shard, self.name)) for shard in
                            self._declared_config['shards']],
             'enable_sharding': [app.Command(self._api.admin.command, {'enableSharding': db_name},
                                             'enabling sharding for database %s' % db_name) for db_name in sharded_dbs],
@@ -77,7 +77,7 @@ class Cluster(object):
 
     @property
     def actual_config(self):
-        logger.debug('Reading configuration of cluster %s', self._name)
+        logger.debug('Reading configuration of cluster %s', self.name)
         db = self._api.config
         out = dict({'databases': {}, 'collections': {}})
         out['mongos'] = [col['_id'] for col in db['mongos'].find()]
@@ -92,11 +92,11 @@ class Cluster(object):
     def check_config(self):
         actual_config = self.actual_config
         if self._declared_config == actual_config:
-            logger.info('Declared configuration of cluster \'%s\' is actual.', self._name)
+            logger.info('Declared configuration of cluster \'%s\' is actual.', self.name)
             return True
         else:
             logger.warning(
                 'Declared configuration of cluster \'%s\' is not actual.\nDeclared:\n---\n%s...\nActual:\n---\n%s...',
-                self._name, yaml.safe_dump(self._declared_config, default_flow_style=False),
+                self.name, yaml.safe_dump(self._declared_config, default_flow_style=False),
                 yaml.safe_dump(actual_config, default_flow_style=False))
             return False