+++ tests/test_cluster.py
@@ -35,7 +35,7 @@ def test_create_objects():
     cluster.Cluster.objects = dict()  # erase objects left from previous test
     assert 2 == cluster.create_objects(('gce-eu', 'aws-va'),
                                        {'gce-eu': sample_cluster_config, 'aws-va': {'mongos': ['l:27017']}})
-    assert [v._name for k, v in cluster.Cluster.objects.items()] == ['gce-eu', 'aws-va']
+    assert [v.name for k, v in cluster.Cluster.objects.items()] == ['gce-eu', 'aws-va']
     assert [v._declared_config for k, v in cluster.Cluster.objects.items()] == [sample_cluster_config,
                                                                                 {'mongos': ['l:27017']}]
 
@@ -43,7 +43,7 @@ def test_create_objects():
 def test_create_objects_with_partial_config():
     cluster.Cluster.objects = dict()  # erase objects left from previous test
     assert 1 == cluster.create_objects(('gce-sc', 'aws-jp'), {'aws-jp': sample_cluster_config})
-    assert [v._name for k, v in cluster.Cluster.objects.items()] == ['aws-jp']
+    assert [v.name for k, v in cluster.Cluster.objects.items()] == ['aws-jp']
     assert [v._declared_config for k, v in cluster.Cluster.objects.items()] == [sample_cluster_config]